backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.14.1
dhanhq==2.0.1
python-dotenv==1.0.1
websockets==13.1
aiohttp==3.10.10
psycopg[binary]==3.2.13
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
email-validator==2.2.0
httpx==0.28.1
pytest==8.3.4
